added Courses

+++ src/main/scala/Human.scala
@@ -1,7 +1,7 @@
 abstract class Human(val name: String = "No Data", val surname: String = "No Data", val age: Int = 0,
-                             val email: String = "No Data", var tokens: Int = 0, var money: Double = 0)
+                             val email: String = "No Data", var tokens: Double = 0, var money: Double = 0)
 {
-  var Tokens: Int = tokens
+  var Tokens: Double = tokens
   private val Money: Double = money
   val role: String = "No Data"
